# openapi.yaml
openapi: 3.1.0
info:
  title: simpletickr API
  version: 0.1.0
  description: Simple portfolio tracker for ETFs, stocks, and crypto.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed
        "400":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /accounts:
    get:
      tags: [Accounts]
      operationId: listAccounts
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get an account by ID
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Accounts]
      operationId: updateAccount
      summary: Update an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Accounts]
      operationId: deleteAccount
      summary: Delete an account
      responses:
        "204":
          description: Account deleted
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Account has linked transactions and cannot be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /portfolios/{portfolioId}/transactions/import/bolero:
    parameters:
      - name: portfolioId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Import]
      operationId: importBoleroTransactions
      summary: Import transactions from a Bolero XLS export
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, accountId]
              properties:
                file:
                  type: string
                  format: binary
                accountId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Import completed (best-effort; check rows for skipped entries)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /import/bolero/analyze:
    post:
      tags: [Import]
      operationId: analyzeBoleroImport
      summary: Parse a Bolero XLS and return unique instruments with existing mapping info
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoleroAnalysisResult"
...
  /data-export:
    get:
      tags: [DataExport]
      operationId: exportData
      summary: Export portfolio data as a JSON file
      description: >
        Exports all portfolios, or only the given portfolios (with only the assets and accounts they reference)
        when portfolioIds is provided.
      parameters:
        - name: portfolioIds
          in: query
          required: false
          explode: true
          schema:
            type: array
            items:
              type: integer
              format: int64
      responses:
        "200":
          description: JSON export file
          content:
            application/json:
              schema:
                type: string
                format: binary
  /data-import:
    post:
      tags: [DataExport]
      operationId: importData
      summary: Import portfolio data from a JSON export file
      parameters:
        - name: dryRun
          in: query
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import analysis (dryRun=true) or import result (dryRun=false)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/DataImportAnalysis"
                  - $ref: "#/components/schemas/DataImportResult"
        "400":
          description: Invalid file or validation errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    CurrentUser:
      type: object
      required:
        - id
        - username
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
    AccountType:
      type: string
      enum:
        - BROKERAGE
        - CRYPTO
        - BANK
        - RETIREMENT
        - OTHER
    Account:
      type: object
      required:
        - id
        - name
        - accountType
        - transactionCount
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "My Fidelity Roth IRA"
        broker:
          type: string
          nullable: true
          example: "Fidelity"
        accountType:
          $ref: "#/components/schemas/AccountType"
        currency:
          type: string
          nullable: true
          example: "USD"
        accountNumber:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
        transactionCount:
          type: integer
          format: int64
          description: "Number of transactions linked to this account"
    AccountRequest:
      type: object
      required:
        - name
        - accountType
      properties:
        name:
          type: string
        broker:
          type: string
          nullable: true
        accountType:
          $ref: "#/components/schemas/AccountType"
        currency:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
...
    BoleroAnalysisResult:
      type: object
      required: [instruments, totalRows, skippedRows]
      properties:
        instruments:
          type: array
          items:
            $ref: "#/components/schemas/BoleroInstrumentInfo"
        totalRows:
          type: integer
          description: Total investment rows found (buys + sells)
        skippedRows:
          type: integer
          description: Non-investment rows (deposits, refunds, parse errors) found during analysis
    BoleroInstrumentInfo:
      type: object
      required: [externalName, rowCount]
      properties:
        externalName:
          type: string
          description: The instrument name as it appears in the Bolero export
        rowCount:
          type: integer
          description: Number of investment transactions for this instrument
        mapping:
          $ref: "#/components/schemas/AssetImportMappingRef"
          nullable: true
    AssetImportMappingRef:
      type: object
      required: [id, assetId]
      properties:
        id:
          type: integer
          format: int64
        assetId:
          type: integer
          format: int64
    ImportResult:
      type: object
      required: [imported, skipped, rows]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRowResult"
    ImportRowResult:
      type: object
      required: [line, status, reason]
      properties:
        line:
          type: integer
        status:
          type: string
          enum: [IMPORTED, SKIPPED]
        reason:
          type: string
...
    DataImportAnalysis:
      type: object
      required:
        - errors
        - assetsToCreate
        - assetsExisting
        - listingsToCreate
        - listingsExisting
        - portfoliosToCreate
        - portfoliosExisting
        - transactionsToImport
        - transactionsSkipped
        - transfersToImport
        - transfersSkipped
      properties:
        errors:
          type: array
          items:
            type: string
        assetsToCreate:
          type: integer
        assetsExisting:
          type: integer
        listingsToCreate:
          type: integer
        listingsExisting:
          type: integer
        portfoliosToCreate:
          type: integer
        portfoliosExisting:
          type: integer
        transactionsToImport:
          type: integer
        transactionsSkipped:
          type: integer
        transfersToImport:
          type: integer
        transfersSkipped:
          type: integer
    DataImportResult:
      type: object
      required:
        - assetsCreated
        - listingsCreated
        - portfoliosCreated
        - transactionsImported
        - transfersImported
      properties:
        assetsCreated:
          type: integer
        listingsCreated:
          type: integer
        portfoliosCreated:
          type: integer
        transactionsImported:
          type: integer
        transfersImported:
          type: integer
...